Create Day_1.py
In the Day-1 of python coding. Here, I written the related code using the print() function. And using print() function how can we print string is describe in brief in this day-1 python code. Like the use of single, double, and triple inverted comma for printing the string or sentences using the print() function in python.