Create Day_2.py
In the Day-2 of Python coding, I describe in details about the print() function that how print() function can print the various of datatypes and to know which datatypes the print() function print, I used type() function inside the print function. So, this is all about the print() function uses and important role in the python. In addition, I also used the comment in the python with "#" to make the python code readable and understandable that what this code is about.